hmm why not use the regular function type?
Ok so the original idea was to be able to compile functions with
functions as arguments. I was noodling around with a new type just to
keep things seperate but I think I can probably use the regular
v-function type in the end.

that's ace as we already have ways of working with them.

Now I'm a bit stuck on how I want to implement this as it essentially
breaks the flow analyzer. We need the whole program to be able to walk
in and out of the lisp functions.

Functions as uniforms is the big issues as that is the case where we
can't get the body code in time.

In that case we should disable flow analysis.

In the default case we need to compile multiple times. Each pass will
find where functions have flow'd to, and then wrap those call sites in
a labels form.

We recompile until all funcall's are gone.

I think this gives us full first class function support..mental

However 'm still at the demoparty so drink be flowin